// front/store/slices/cartSlice.ts
//front/store/slices/cartSlice.ts

import { CartItem, CartState } from '@/types/cart.types';
import { Product } from '@/types/product.types';
import { createSlice, PayloadAction } from '@reduxjs/toolkit';

const initialState: CartState = {
    items: [],
    totalItems: 0,
    totalPrice: 0,
};

const calculateTotals = (items: CartItem[]) => {
    const totalItems = items.reduce((sum, item) => sum + item.quantity, 0);
    const totalPrice = items.reduce((sum, item) => sum + item.product.price * item.quantity, 0);

    return { totalItems, totalPrice };
}

const cartSlice = createSlice({
    name: "cart",
    initialState,
    reducers: {
        addToCart: (state, action: PayloadAction<Product>) => {
            const existingItem = state.items.find(
                (item) => item.product.id === action.payload.id
            )

            if (existingItem) {
                existingItem.quantity += 1
            } else {
                state.items.push({
                    product: action.payload,
                    quantity: 1,
                    price: action.payload.price,
                    id: crypto.randomUUID(),
                    cartId: "",
                    productId: action.payload.id,
                    createdAt: new Date().toISOString(),
                    updatedAt: new Date().toISOString(),
                })
            }

            const totals = calculateTotals(state.items)
            state.totalItems = totals.totalItems;
            state.totalPrice = totals.totalPrice;
        },

        decrementQuantity: (state, action: PayloadAction<string>) => {
            const item = state.items.find(
                (item) => item.product.id === action.payload
            );

            if (item) {
                if (item.quantity <= 1) {
                    state.items = state.items.filter(
                        (item) => item.product.id !== action.payload
                    )
                } else {
                    item.quantity -= 1
                }

                const totals = calculateTotals(state.items)
                state.totalItems = totals.totalItems;
                state.totalPrice = totals.totalPrice;
            }
        },
        incrementQuantity: (state, action: PayloadAction<string>) => {
            const item = state.items.find(
                (item) => item.product.id === action.payload
            )

            if (item) {
                item.quantity += 1;

                const totals = calculateTotals(state.items)
                state.totalItems = totals.totalItems;
                state.totalPrice = totals.totalPrice;
            }
        },
        removeFromCart: (state, action: PayloadAction<string>) => {
            state.items = state.items.filter(
                (item) => item.product.id !== action.payload
            );

            const totals = calculateTotals(state.items)
            state.totalItems = totals.totalItems;
            state.totalPrice = totals.totalPrice;
        },
        clearAllCart: (state) => {
            state.items = [];
            state.totalItems = 0;
            state.totalPrice = 0;
        }
    },
});

export const { addToCart, decrementQuantity, incrementQuantity, removeFromCart, clearAllCart } = cartSlice.actions;
export default cartSlice.reducer;